캐릭터, 패시브템, 무기 파라미터 추가 상점, 캐선, 무기선택에서도 파리미터 보여주기 상점에서 무기구매하면 스탯반영 및 스탯창 갱신

+++ Client/CPlayer.h
@@ -32,7 +32,7 @@ struct playerParameter {
     float			m_fMeleeCoef;		//근거리 최종뎀 %
     float			m_fRangeCoef;		//원거리 최종뎀 %
     float			m_fAttackSpeedCoef; // 공격속도 %
-    float			m_iCriticalAcc;		// 크확 %
+    int			    m_iCriticalAcc;		// 크확 %
     float			m_fSpeed;			// 속도 계수.
 };
 class CPlayer :
